pyproject.toml
[tool.poetry]
name = "ml-sandbox"
version = "0.1.0"
description = "A playground and workspace for experiment with deep learning models."
authors = ["Sidharth Baskaran <62273881+sidnb13@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ml_sandbox"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
wandb = "^0.15.8"
absl-py = "^1.4.0"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
ml-collections = "^0.1.1"
matplotlib = "^3.7.2"
numpy = "^1.25.2"
tiktoken = "^0.4.0"
colorama = "^0.4.6"
scipy = "^1.11.2"
jaxlib = "^0.4.14"
jax = {extras = ["cpu"], version = "^0.4.14"}
flax = {git = "https://github.com/google/flax.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
